Change dtos namespaces
